src/main/java/br/com/alura/codechella/domain/entities/users/Users.java
package br.com.alura.codechella.domain.entities.users;

import java.time.LocalDate;

/**
 * In the Domain layer is to define all the business rules.
 * Pure Java Code and totally testable.
 */
public class Users {

    private String cpf;
    private String nome;
    private LocalDate nascimento;
    private String email;

    public Users(String cpf, String nome, LocalDate nascimento, String email) {
        if (cpf == null || !cpf.matches("\\d{3}\\.\\d{3}\\.\\d{3}\\-\\d{2}")) {
            throw new IllegalArgumentException("CPF is incorret");
        }
        this.cpf = cpf;
        this.nome = nome;
        this.nascimento = nascimento;
        this.email = email;
    }

    public String getCpf() {
        return cpf;
    }

    public void setCpf(String cpf) {
        this.cpf = cpf;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public LocalDate getNascimento() {
        return nascimento;
    }

    public void setNascimento(LocalDate nascimento) {
        this.nascimento = nascimento;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}
